requirements.txt
# FastAPI and ASGI server
fastapi>=0.115.7
uvicorn[standard]>=0.27.0
gunicorn>=21.2.0

# Authentication and security
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.9

# Environment and configuration
python-dotenv>=1.0.0
pydantic>=2.6.0
pydantic-settings>=2.1.0

# Google Gemini API
google-generativeai>=0.3.2
google-cloud-logging>=3.9.0

# Monitoring and metrics
prometheus-client>=0.19.0
opentelemetry-api>=1.23.0
opentelemetry-sdk>=1.23.0
opentelemetry-instrumentation-fastapi>=0.44b0

# Testing (optional in production)
pytest>=8.0.0
pytest-cov>=4.1.0
httpx>=0.26.0

# Production utilities
structlog>=24.1.0
python-json-logger>=2.0.7
tenacity>=8.2.3 